cleaned up root directory by adding new thematic modules or moving files in existing ones
