+++ src/pipeline/steps/clustering_util.py
@@ -3,7 +3,7 @@ from typing import Optional, Union, Callable
 import flask
 from werkzeug.datastructures import FileStorage
 
-from load_utils import FactoryLoader
+from src.pipeline.load_utils import FactoryLoader
 from main_utils import StepsName, BaseUtil
 from src.core.cluster_functions import PhraseClusterFactory
