+++ src/pipeline/steps/embedding_util.py
@@ -7,7 +7,7 @@ from main_utils import StepsName, BaseUtil
 from src.storage.marqo_external_utils import MarqoEmbeddingStore
 from src.core.embedding_functions import SentenceEmbeddingsFactory
 from src.common.util_functions import load_pickle
-from load_utils import FactoryLoader
+from src.pipeline.load_utils import FactoryLoader
 
 
 # ToDo: I need to check the storage_method config/kwarg; it works but there are some redundancies
