+++ test/test_document_clustering_on_corpus.py
@@ -3,7 +3,7 @@ import unittest
 
 from sklearn.cluster import KMeans
 
-from load_utils import FactoryLoader
+from src.pipeline.load_utils import FactoryLoader
 from main_utils import StepsName
 from src.core.cluster_functions import WordEmbeddingClustering
